Updated UI and contact page

+++ admin/view_crop.php
@@ -25,39 +25,57 @@ $result = mysqli_query($conn, "SELECT * FROM crops");
 <html>
 <head>
 <title> View Crops </title>
-<link rel ="stylesheet" href ="../assests/css/style1.css">
+<link rel ="stylesheet" href ="../assests/css/style_admin.css">
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
 </head>
 
 <body>
 
-<h2> Crop list </h2>
+<div class="view-hero">
+        
+        <div class="table-box">
+            
+            <h1 class="page-title">Crops Information</h1>
 
-<table border="1" cellpadding="10">
+            <table class="crop-table">
+                <thead>
+                    <tr>
+                        <th>Crop ID</th>
+                        <th>Crop Name</th>
+                        <th>Season</th>
+                        <th>Description</th>
+                        <th>Image</th>
+                        <th>Action</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    <?php while ($row = mysqli_fetch_assoc($result)) { ?>
+                    <tr>
+                        <td> <?php echo $row['crop_id']; ?></td>
+						
+						<td> <?php echo $row['crop_name']; ?></td>
+                        
+                        <td> <?php echo $row['season']; ?></td>
+                        
+                        <td style="max-width: 300px;"> <?php echo $row['description']; ?></td>
 
-<tr>
-<th>Crop Id</th>
-<th>Crop Name</th>
-<th>Season</th>
-<th>Description</th>
-<th>Action</th>
-</tr>
+                        <td> <img src="<?php echo $row['crop_image']; ?>" class="table-img" alt="Crop"  height="15px"width="100%"> </td>
 
-<?php while ($row = mysqli_fetch_assoc($result)) { ?>
-	
-<tr>
-<td> <?php echo $row['crop_id'] ?> </td>
-<td> <?php echo $row['crop_name'] ?> </td>
-<td> <?php echo $row['season'] ?> </td>
-<td> <?php echo $row['description'] ?> </td>
-<td>
-<a href="view_crop.php?delete=<?php echo $row['crop_id'];?>"
-onclick = "return confirm('Are you sure?')"> Delete
-</td>
-</tr>
+                        <td> <a href="view_crop.php?delete=<?php echo $row['crop_id']; ?>" class="delete-btn" onclick="return confirm('Are you sure you want to delete this crop?')">
+                             <i class="fas fa-trash"></i>
+                            </a>
+                        </td>
+                    </tr>
+                    <?php 
+                    } 
+                    ?>
+                </tbody>
+            </table>
 
-<?php } ?>
-</table>
+        </div>
 
-<p> <a href="dashboard.php"> Back to dashboard</a></p>
+        <a href="dashboard.php" style="color: white; text-decoration: none; font-size: 1.1rem;"> Back to dashboard </a>
+
+    </div>
 </body>
 </html>
